requirements.txt
flask
pillow
numpy
tensorflow==2.13.0